+++ CHANGELOG.md
@@ -56,6 +56,7 @@ All notable changes to this project will be documented in this file.
 - lost password using supabase auth
 - reset password using supabase auth
 - update email using supabase auth
+- update password using supabase auth
 
 #### Supabase